--- src/components/feedbackboard/gradientBoard/GradientBoard.module.css
@@ -1,47 +1,0 @@
-h1,
-h2 {
-  margin: 0;
-}
-.gradientBoard {
-  background-image: url(/assets/suggestions/mobile/background-header.png);
-  background-repeat: no-repeat;
-  background-size: cover;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem 1rem 0.94rem 1rem;
-}
-h1 {
-  color: rgba(var(--text-white));
-  font-size: 15px;
-  font-weight: 700;
-  letter-spacing: -0.187px;
-}
-h2 {
-  font-size: 13px;
-  font-weight: 500;
-  color: rgba(var(--text-white));
-  opacity: 0.75;
-}
-@media screen and (min-width: 48rem) {
-  .gradientBoard {
-    background-image: url(/assets/suggestions/tablet/background-header.png);
-    width: 13.9375rem;
-    height: 11.125rem;
-    border-radius: 0.625rem;
-    align-items: flex-end;
-  }
-  h2 {
-    margin-top: 0.5rem;
-  }
-  .menuIcon {
-    display: none;
-  }
-  .tabletBar {
-    margin: 2.5rem auto 0 auto;
-    gap: 1rem;
-    width: 75%;
-    display: flex;
-    justify-content: center;
-  }
-}
