+++ src/components/feedbackboard/headerContainer/MobileHeader.jsx
@@ -4,8 +4,7 @@ import GradientBoard from "../gradientBoard/GradientBoard";
 import suggestion from "/assets/suggestions/icon-suggestions.svg";
 export default function MobileHeader() {
   return (
-    <div className="flex w-full flex-col lg:hidden">
-      <GradientBoard />
+    <div className="flex w-full flex-col lg:flex-col">
       <div className={styles.blueCont}>
         <div className={styles.headerSuggestions}>
           <img className={styles.suggestion} src={suggestion} alt="" />
