+++ src/components/feedbackboard/headerContainer/MobileHeader.module.css
@@ -65,3 +65,13 @@ select {
     letter-spacing: -0.01563rem;
   }
 }
+@media screen and (min-width: 90rem) {
+  .blueCont {
+    display: flex;
+    margin: 30px 6rem 1.5rem auto;
+    width: 100%;
+    max-width: 58.75rem;
+    border-radius: 0.625rem;
+    justify-content: space-between;
+  }
+}
