+++ src/components/feedbackboard/sideBar/SideBar.jsx
@@ -1,7 +1,6 @@
 import ButtonBoard from "./buttonBoard/ButtonBoard";
 import styles from "./SideBar.module.css";
 import RoadBoard from "./roadBoard/RoadBoard";
-import GradientBoard from "../gradientBoard/GradientBoard";
 
 export default function SideBar({ menu, toggleMenu }) {
   return (
